Optimise rest api loading
