+++ src/Domain/Services/GoogleAnalytics.php
@@ -22,13 +22,12 @@ class GoogleAnalytics {
 	 */
 	public function __construct( AssetApi $asset_api ) {
 		$this->asset_api = $asset_api;
-		$this->init();
 	}
 
 	/**
 	 * Hook into WP.
 	 */
-	protected function init() {
+	public function init() {
 		add_action( 'init', array( $this, 'register_assets' ) );
 		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_scripts' ) );
 		add_filter( 'script_loader_tag', array( $this, 'async_script_loader_tags' ), 10, 3 );
